+++ src/main/java/BackERP/repository/RepositoryInventario.java
@@ -19,5 +19,5 @@ public interface RepositoryInventario extends JpaRepository<erpInventario, Long>
 
     // Nuevo metodo para buscar inventario por producto y ubicación
 
-    erpInventario findByIdProducto_IdProductoAndIdUbicacion(int idProducto, int idUbicacion);
+    erpInventario findByIdProducto_IdProductoAndIdUbicacion(Long idProducto, int idUbicacion);
 }
